updates ConfigurationEntryCreator class as ConfigurationEntryCollection
also adds package-private methods for mocking addition of ConfigurationEntry instances and adds JavaDocs
